Reorganize and write proper doc
