+++ .spyproject/config/backups/codestyle.ini.bak
@@ -1,0 +1,8 @@
+[codestyle]
+indentation = True
+edge_line = True
+edge_line_columns = 79
+
+[main]
+version = 0.2.0
+
