+++ .spyproject/config/backups/encoding.ini.bak
@@ -1,0 +1,6 @@
+[encoding]
+text_encoding = utf-8
+
+[main]
+version = 0.2.0
+
