+++ .spyproject/config/backups/workspace.ini.bak
@@ -1,0 +1,12 @@
+[workspace]
+restore_data_on_startup = True
+save_data_on_exit = True
+save_history = True
+save_non_project_files = False
+project_type = 'empty-project-type'
+recent_files = ['clock.py']
+
+[main]
+version = 0.2.0
+recent_files = []
+
