+++ .spyproject/config/defaults/defaults-codestyle-0.2.0.ini
@@ -1,0 +1,5 @@
+[codestyle]
+indentation = True
+edge_line = True
+edge_line_columns = 79
+
