+++ .spyproject/config/defaults/defaults-encoding-0.2.0.ini
@@ -1,0 +1,3 @@
+[encoding]
+text_encoding = utf-8
+
