+++ .spyproject/config/defaults/defaults-vcs-0.2.0.ini
@@ -1,0 +1,4 @@
+[vcs]
+use_version_control = False
+version_control_system = 
+
