+++ .spyproject/config/defaults/defaults-workspace-0.2.0.ini
@@ -1,0 +1,6 @@
+[workspace]
+restore_data_on_startup = True
+save_data_on_exit = True
+save_history = True
+save_non_project_files = False
+
