+++ .spyproject/config/vcs.ini
@@ -1,0 +1,7 @@
+[vcs]
+use_version_control = False
+version_control_system = 
+
+[main]
+version = 0.2.0
+
